refactor(blockchain): extract block_builder module
Move build_block and its helpers out of store.rs into a dedicated
block_builder module: PostBlockCheckpoints, build_block (the public
entry point at the top of the file), select_attestations,
pick_best_candidate, ChainContext, ProjectedState, entry_passes_filters,
score_entry, EntryScore, build_running_votes, compact_attestations,
extend_proofs_greedily, union_aggregation_bits, is_genesis_self_vote,
trace_skipped_attestation, plus the related tests.

store.rs now imports build_block + PostBlockCheckpoints from
block_builder and keeps the on-block import path, attestation
validation, verify_block_signatures, and the live store actor.
